revset: move primitive typed expression helpers to parser module
